requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
requests>=2.30.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorboard>=2.13.0
wandb